fix: Remove exposure TODOs (#2177)
* fix: Remove exposure TODOs

* fix: Make `exposure` a Double

+++ package/android/src/main/java/com/mrousavy/camera/CameraView.kt
@@ -78,7 +78,7 @@ class CameraView(context: Context) :
   var isActive = false
   var torch: Torch = Torch.OFF
   var zoom: Float = 1f // in "factor"
-  var exposure: Float = 1f // TODO: Implement exposure bias
+  var exposure: Double = 1.0
   var orientation: Orientation = Orientation.PORTRAIT
   var enableZoomGesture: Boolean = false
     set(value) {
